Синхронизация потоков в критической секции
